+++ include/podio/CollectionIDTable.h
@@ -5,6 +5,7 @@
 #include <mutex>
 #include <string>
 #include <vector>
+#include <cstdint>
 
 namespace podio {
 
@@ -20,15 +21,15 @@ public:
   CollectionIDTable& operator=(CollectionIDTable&&) = default;
 
   /// constructor from existing ID:name mapping
-  CollectionIDTable(std::vector<int>&& ids, std::vector<std::string>&& names);
+  CollectionIDTable(std::vector<uint64_t>&& ids, std::vector<std::string>&& names);
 
-  CollectionIDTable(const std::vector<int>& ids, const std::vector<std::string>& names);
+  CollectionIDTable(const std::vector<uint64_t>& ids, const std::vector<std::string>& names);
 
   /// return collection ID for given name
-  int collectionID(const std::string& name) const;
+  uint64_t collectionID(const std::string& name) const;
 
   /// return name for given collection ID
-  const std::string name(int collectionID) const;
+  const std::string name(uint64_t collectionID) const;
 
   /// Check if collection name is known
   bool present(const std::string& name) const;
@@ -39,13 +40,13 @@ public:
   };
 
   /// return the ids
-  const std::vector<int>& ids() const {
+  const std::vector<uint64_t>& ids() const {
     return m_collectionIDs;
   }
 
   /// register new name to the table
   /// returns assigned collection ID
-  int add(const std::string& name);
+  uint64_t add(const std::string& name);
 
   /// Prints collection information
   void print() const;
@@ -56,7 +57,7 @@ public:
   }
 
 private:
-  std::vector<int> m_collectionIDs{};
+  std::vector<uint64_t> m_collectionIDs{};
   std::vector<std::string> m_names{};
   mutable std::unique_ptr<std::mutex> m_mutex{nullptr};
 };
