+++ src/CollectionIDTable.cc
@@ -10,22 +10,22 @@ namespace podio {
 CollectionIDTable::CollectionIDTable() : m_mutex(std::make_unique<std::mutex>()) {
 }
 
-CollectionIDTable::CollectionIDTable(std::vector<int>&& ids, std::vector<std::string>&& names) :
+CollectionIDTable::CollectionIDTable(std::vector<uint64_t>&& ids, std::vector<std::string>&& names) :
     m_collectionIDs(std::move(ids)), m_names(std::move(names)), m_mutex(std::make_unique<std::mutex>()) {
 }
 
-CollectionIDTable::CollectionIDTable(const std::vector<int>& ids, const std::vector<std::string>& names) :
+CollectionIDTable::CollectionIDTable(const std::vector<uint64_t>& ids, const std::vector<std::string>& names) :
     m_collectionIDs(ids), m_names(names), m_mutex(std::make_unique<std::mutex>()) {
 }
 
-const std::string CollectionIDTable::name(int ID) const {
+const std::string CollectionIDTable::name(uint64_t ID) const {
   std::lock_guard<std::mutex> lock(*m_mutex);
   const auto result = std::find(begin(m_collectionIDs), end(m_collectionIDs), ID);
   const auto index = std::distance(m_collectionIDs.begin(), result);
   return m_names[index];
 }
 
-int CollectionIDTable::collectionID(const std::string& name) const {
+uint64_t CollectionIDTable::collectionID(const std::string& name) const {
   std::lock_guard<std::mutex> lock(*m_mutex);
   const auto result = std::find(begin(m_names), end(m_names), name);
   const auto index = std::distance(m_names.begin(), result);
@@ -46,7 +46,7 @@ bool CollectionIDTable::present(const std::string& name) const {
   return result != end(m_names);
 }
 
-int CollectionIDTable::add(const std::string& name) {
+uint64_t CollectionIDTable::add(const std::string& name) {
   std::lock_guard<std::mutex> lock(*m_mutex);
   const auto result = std::find(begin(m_names), end(m_names), name);
   uint64_t ID = 0;
